Uni additions

+++ Assets/Player/Player.cs
@@ -17,7 +17,7 @@ namespace TheWorkforce
         protected PlayerController _controller { get; private set; }
 
         //public Player(PlayerController playerController, SlotCollection inventory, Toolbelt toolbelt, Movement movement)
-        public Player(PlayerController playerController, SlotCollection inventory, Movement movement) : base(playerController.Id, null)
+        public Player(PlayerController playerController, SlotCollection inventory, Movement movement) : base((uint)playerController.Id, null)
         {
             _controller = playerController;
             Inventory = inventory;
